monorepo/pyproject.toml
[tool.poetry]
name = "tlt"
version = "0.1.0"
description = ""
authors = ["David Miura"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.0"
"discord.py" = "^2.3.2"
fastapi = "^0.109.0"
uvicorn = "^0.27.0"
fastmcp = "^2.10.5"
httpx = "^0.28.1"
casbin = "^1.17.0"
langchain = "^0.3.26"
langchain-openai = "^0.3.27"
langchain-core = "^0.3.68"
langgraph = "^0.5.2"
pillow = "^10.0.0"
requests = "^2.31.0"
mcp-server = "^0.1.4"
loguru = "^0.7.0"
aiofiles = "^24.1.0"
aiohttp = "^3.10.0"
pillow-avif-plugin = "^1.5.2"
pytesseract = "^0.3.13"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
